Eatfit (#35)
Eatfit page now allows users to control their percentage of fats, carbs and proteins for their diet. In addition, the users total suggested calorie intake for the day is provided and split into 3 meals: Breakfast (20%), Lunch (35%) and dinner (45%).

In the future it would be nice to allow the user to control their daily percentage intake for their meals and give them the option to add and delete meals and assign descriptions to these.